requirements.txt

pandas==1.3.5
matplotlib==3.5.1
statsmodels==0.13.2
pyarrow==8.0.0
pyyaml==6.0
pathlib==1.0.1
click==8.1.3
argparse==1.4.0
scikit-learn-1.3.2